requirements.txt
fastapi
uvicorn
pymongo
chromadb
beautifulsoup4
requests
langchain
sentence-transformers
python-dotenv
google-generativeai
tiktoken